requirements.txt
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.2.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cron-descriptor==1.4.3
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
h11==0.14.0
idna==3.8
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
lockfile==0.12.2
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.1.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.1
requests==2.32.3
requests-futures==1.0.1
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
urllib3==2.2.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
zipp==3.17.0
